+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
         - Use `verify_sort` when obtaining borders.
         - Support for penalizing the optimization results using borderlines. The algorithm returns `PENALTY * d`, where `d` is the distance to the nearest borderline point that the candidate point belongs to.
         - Parameter `use_borderlines` that activates borderline-based penalization.
+- Penalizers
+    - New group of algorithms which are used for evaluating whether the candidate is incorrect. In that case penalty is computed.
+    - _Count_
+        - Penalizer 'Count' is the basic penalizer used in ng_trajectory. It counts number of points outside the valid area.
 - Segmentators
     - _Flood Fill_
         - Function `segmentDistance()` that computes distance from a point to a segment.